fix warnings
svn path=/branches/extragear/kde3/graphics/digikam/; revision=735592

+++ libs/dimg/dimg.cpp
@@ -1340,7 +1340,7 @@ void DImg::rotate(ANGLE angle)
         uint w  = width();
         uint h  = height();
 
-        uint middle_line = -1;
+        int middle_line = -1;
         if (h % 2)
            middle_line = h / 2;
 
@@ -1365,7 +1365,7 @@ void DImg::rotate(ANGLE angle)
 
                     line1++;
                     line2--;
-                    if (y == middle_line && x * 2 >= w)
+                    if ((int)y == middle_line && x * 2 >= w)
                         break;
                 }
             }
@@ -1392,8 +1392,8 @@ void DImg::rotate(ANGLE angle)
 
                     line1++;
                     line2--;
-                    if (y == middle_line && x * 2 >= w)
-			break;
+                    if ((int)y == middle_line && x * 2 >= w)
+                        break;
                 }
             }
         }
